// types.ts
export interface Student {
  id: string; // Student Number MDS1...
  name: string;
  class: string;
  gender: 'Male' | 'Female' | 'Other';
  dob: string;
  dateOfAdmission: string;
  club: string;
  guardianName: string;
  residentialAddress: string;
  contact: string;
  status: 'Active' | 'Alumni' | 'Transferred' | 'Dropped out';
  photo?: string;
}

export interface FeeStructure {
  id: string;
  className: string;
  tuitionFee: number;
  developmentFee: number;
  examFee: number;
  otherFees: number;
  term: string;
  year: string;
}

export interface FeePayment {
  id: string;
  studentId: string;
  amount: number;
  date: string;
  paymentMethod: 'Cash' | 'Bank Transfer' | 'Mobile Money';
  receiptNumber: string;
  term: string;
  year: string;
  recordedBy: string;
  note?: string;
}
// ...
export interface ClassInfo {
  id: string;
  name: string; // e.g. "Grade 10 A"
  gradeTeacherId: string; // Teacher ID
  subjectIds: string[]; // List of subjects taught in this class
}
// ...
export type UserRole = 'ADMIN' | 'TEACHER' | 'STUDENT' | 'ATTENDANCE_OFFICER';